reorg: move core modules to engine/ and normalize binary/cli loading
