+++ src/bin/timelapse.rs
@@ -1,6 +1,3 @@
-#[path = "../cli.rs"]
-mod cli;
-
 fn main() -> anyhow::Result<()> {
-    cli::run()
+    timelapse::cli::run()
 }
